fix : cleaned index.vue and added test.vue to test our implemented functionalities (#16)
